Refactor ProcessWithResult class.
